+++ ethexe/common/src/lib.rs
@@ -30,23 +30,20 @@ pub mod db;
 pub mod events;
 pub mod gear;
 pub mod tx_pool;
+pub mod utils;
 
 pub use crypto::*;
 pub use gear_core;
 pub use gprimitives;
+pub use k256;
 pub use sha3;
+pub use utils::*;
 
 use alloc::vec::Vec;
 use db::BlockHeader;
 use events::BlockEvent;
 use gprimitives::H256;
 use parity_scale_codec::{Decode, Encode};
-
-pub const fn u64_into_uint48_be_bytes_lossy(val: u64) -> [u8; 6] {
-    let [_, _, b1, b2, b3, b4, b5, b6] = val.to_be_bytes();
-
-    [b1, b2, b3, b4, b5, b6]
-}
 
 // TODO #4547: move to submodule.
 #[derive(Debug, Clone, PartialEq, Eq)]
