+++ ethexe/common/src/utils.rs
@@ -1,0 +1,18 @@
+/// Decodes hexed string to a byte array.
+pub fn decode_to_array<const N: usize>(s: &str) -> Result<[u8; N], hex::FromHexError> {
+    // Strip the "0x" prefix if it exists.
+    let stripped = s.strip_prefix("0x").unwrap_or(s);
+
+    // Decode
+    let mut buf = [0u8; N];
+    hex::decode_to_slice(stripped, &mut buf)?;
+
+    Ok(buf)
+}
+
+/// Converts u64 to a 48-bit unsigned integer, represented as a byte array in big-endian order.
+pub const fn u64_into_uint48_be_bytes_lossy(val: u64) -> [u8; 6] {
+    let [_, _, b1, b2, b3, b4, b5, b6] = val.to_be_bytes();
+
+    [b1, b2, b3, b4, b5, b6]
+}
